requirements.txt
fastapi
uvicorn
pydantic
requests
httpx
pytest
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib